+++ src/app/search/maps/page.tsx
@@ -1,9 +1,6 @@
+import { SearchMapsPage } from "@/layers/search/page/maps/page";
 import React from "react";
 
 export default function Page() {
-  return (
-    <section className="flex gap-2 bg-accent/50 h-full flex-grow rounded-xl justify-center items-center text-center">
-      <p>Seleccionar una sucursal para ver sus detalles.</p>
-    </section>
-  );
+  return <SearchMapsPage />;
 }
