+++ src/app/search/page.tsx
@@ -1,8 +1,4 @@
-import HomeHeader from "@/layers/home/components/home-header";
-import CategoryFilter from "@/layers/search/components/category-filter";
-import LocationFilter from "@/layers/search/components/location-filter";
-import Search from "@/layers/search/components/search";
-import SearchTable from "@/layers/search/components/table";
+import { SearchPage } from "@/layers/search/page";
 import React, { Suspense } from "react";
 
 interface PageProps {
@@ -15,49 +11,9 @@ interface PageProps {
 }
 
 export default function Page({ searchParams }: PageProps) {
-  const query = searchParams?.query || "";
-  const category = searchParams?.category || "";
-  const city = searchParams?.city || "";
-  const currentPage = Number(searchParams?.page) || 1;
-
   return (
-    <div className="flex flex-col   bg-muted">
-      <HomeHeader>
-        <h1 className="text-lg font-light ">Encuentra servicios cerca de ti</h1>
-      </HomeHeader>
-
-      <main className=" h-screen">
-        <div className="container mx-auto px-4 py-8 h-full ">
-          <div className="flex space-x-4 mb-8">
-            <div className="flex-grow">
-              <Search placeholder="Buscar servicios, negocios o tratamientos" />
-            </div>
-            <button className="bg-green-600 text-white px-6 py-2 rounded-md hover:bg-green-700">
-              Buscar
-            </button>
-          </div>
-
-          <div className="flex space-x-8">
-            <aside className="w-1/4 space-y-4">
-              <h2 className="text-xl font-semibold mb-4">Filtros</h2>
-              <LocationFilter />
-              <CategoryFilter />
-            </aside>
-
-            <div className="w-3/4 ">
-              <Suspense fallback={<div>Cargando resultados...</div>}>
-                <SearchTable
-                  view="grid"
-                  city={city}
-                  query={query}
-                  currentPage={currentPage}
-                  category={category}
-                />
-              </Suspense>
-            </div>
-          </div>
-        </div>
-      </main>
-    </div>
+    <Suspense fallback={"Loading .."}>
+      <SearchPage searchParams={searchParams} />
+    </Suspense>
   );
 }
