+++ src/layers/search/page/maps/page/index.ts
@@ -1,0 +1,1 @@
+export * from "./map-page";
