+++ src/layers/search/page/maps/page/map-page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function SearchMapsPage() {
+  return (
+    <section className="flex gap-2 bg-accent/50 h-full flex-grow rounded-xl justify-center items-center text-center">
+      <p>Seleccionar una sucursal para ver sus detalles.</p>
+    </section>
+  );
+}
